Fixed precision bug on MinGW.

+++ src/Renderers/DepthComplexityRenderer.cpp
@@ -251,7 +251,7 @@ void DepthComplexityRenderer::renderGui() {
             sgl::Color newColor = sgl::colorFromFloat(colorSelection.x, colorSelection.y, colorSelection.z, 1.0f);
             renderColor = newColor;
             intensity = 0.0001f + 3*colorSelection.w;
-            numFragmentsMaxColor = std::max(maxComplexity, 4ul)/intensity;
+            numFragmentsMaxColor = std::max(maxComplexity, uint64_t(4ull))/intensity;
             reRender = true;
         }
     }
@@ -284,7 +284,7 @@ void DepthComplexityRenderer::computeStatistics(bool isReRender) {
     uint64_t usedLocations = 0;
     uint64_t maxComplexity = 0;
     uint64_t minComplexity = 0;
-#pragma omp parallel for reduction(+:totalNumFragments,usedLocations) reduction(max:maxComplexity) reduction(min:minComplexity) schedule(static)
+    #pragma omp parallel for reduction(+:totalNumFragments,usedLocations) reduction(max:maxComplexity) reduction(min:minComplexity) schedule(static)
     for (int i = 0; i < bufferSize; i++) {
         totalNumFragments += data[i];
         if (data[i] > 0) {
@@ -305,7 +305,7 @@ void DepthComplexityRenderer::computeStatistics(bool isReRender) {
         if (!isReRender) {
             firstFrame = false;
         }
-        numFragmentsMaxColor = std::max(maxComplexity, 4ul)/intensity;
+        numFragmentsMaxColor = std::max(maxComplexity, uint64_t(4ull))/intensity;
     }
 
     if (performanceMeasureMode) {
